Mod usu

+++ pet/src/main/java/com/pethistory/pet/dtos/DtoUsuario.java
@@ -39,7 +39,7 @@ public class DtoUsuario {
     private Date fechaNac;
 
     @NotNull
-    private Integer tipdoc;
+    private Long tipdoc;
 
     @NotNull
     private String nomTipoDoc;
